Moving towards a more flexible Entry/Column representation.
We'd like to be able to let users specify their own column names and
types, and enter different fields for entries. For example, I'd like a
"year of birth" field for tennis entries which I could then group on,
much like the country group.

For now we just refactor country group to be an instance of a new
class "property group," and make the entry type abstract.